[tasklist] Do not lose the pinned status when dragging a window
When dragging and dropping a window to reorder the tasklist, if the
window was pinned, then it will get unpinned, which is wrong.

Of course, this is not perfect since placing this pinned window in a
specific spot in a workspace will also change its sort order on other
workspaces. But there's no way to make this work correctly: if a user
wants to have a pinned window at a specific spot in his tasklist, this
just cannot work in all different workspaces at the same time.

https://bugzilla.gnome.org/show_bug.cgi?id=603516