[tasklist] Fix a GSeal issue
We were using the event_window field of GtkButton, which is sealed with
no accessor (that makes sense). What we're interested in is the position
of the GtkButton on the screen, and it's easy to find it another way.

There's just a small trick: the window of a GtkButton is actually the
window of its parent; once we know this, we just need to use the
allocation information to get the relative position and we can deduce
the absolution position of the widget.