tasklist: Reorganize how the glow button works
Instead of screenshotting twice and blending the two screenshots in a
timeout handler, we now do the blending in the draw handler and only
compute the glow factor in the timeout handler.

This way we ensure that drawing only happens in the draw handler.
Also, the code looks much nicer now.