core: Add wnck_shutdown()
Add wnck_shutdown() method that stops listening to events and tear down all
resources from libwnck. This should be done if you are not going to need the
state change notifications for an extended period of time, to avoid wakeups
with every key and focus event. After this all Wnck object references you might
still hold are invalid.

https://bugzilla.gnome.org/show_bug.cgi?id=642692