TaskList: use new Startup Notification API
The startup-notification API was recently fixed to use pointers to 'time_t'
instead of pointers to 'long' in the functions
sn_startup_sequence_get_initiated_time() and
sn_startup_sequence_get_last_active_time(). See
http://cgit.freedesktop.org/startup-notification/commit/?id=a7e49fefde18ea8d5bada8096d32f23bcfb5a6dc
This is necessary for correct operation on 32bit platforms that use 64bit
time_t (e.g. OpenBSD >= 5.5, Linux x32 ABI).

Unfortunately, it is unknown if or when a new startup-notification release
containing this patch will be made. Until this situation clears up, downstream
distributions might or might not choose to apply the above patch to
startup-notification.

If startup-notification is patched, or a new release containing the patch is
made, then libwnck needs the attached patch to avoid corruption of stack
variables on 32bit platforms if startup-notifications are enabled.