Add wrap-on-scroll option for WnckPager widget
This commit is a minor and optional enhancement of the WnckPager
widget.

It allows to scroll over the borders of workspace boundaries.

This behavior, which is disabled by default, might be enabled with
wnck_pager_set_wrap_on_scroll. When activated, a scroll event which
would otherwise do nothing, i.e. over the border, will now scroll from
the first to the last workspace and from the last to the first. With
this option disabled the behavior is exactly the same as before.

https://bugzilla.gnome.org/show_bug.cgi?id=780490