tasklist: Add signals for enter and leave notify events
Panels implementing applets with tasklists may need access to
enter-notify-event and leave-notify-event on the individual task
widgets, so that actions like displaying thumbnails of windows can
be implemented.

On event firing, the tasklist will emit a new signal and send a
GList of all WnckWindows that belong to the task.