build: Fix build with meson 0.50
With meson 0.50, we no longer just can define an empty Install_dir to specify the
file should not be installed, but we need to specify true/false in Install. Otherwise,
the build fails with:

DEPRECATION: install_dir: kwarg must be a string and not an empty array. Please use the install: kwarg to enable or disable installation. This will be a hard error in the next release.

meson.build:87:2: ERROR: "install_dir" must be specified when "install" in a configure_file is true