class-group: Properly weak_unref window
A weak_ref is added in `_wnck_class_group_add_window` but it isn't
weak_unref in `_wnck_class_group_remove_window`. This means that when
the class group is destroyed, and then the window is destroyed (such as
the user code holding a strong reference to that window), the handler
would operate on an already-destroyed class group.